requirements.txt
Django>=2.2.10,<2.3
psycopg2>=2.8,<2.9
djangorestframework>=3.9.4,<3.10
django-filter>=2.0,<2.1
boto3>=1.7,<1.8
astropy>=4.0,<4.1
django-crispy-forms>=1.7,<1.8
markdown>=2.6,<2.7
gunicorn[gevent]==19.9.0
django-cors-headers>=2.4,<2.5
python3-memcached>=1.51,<2.0
kombu>=4.6,<4.7
ocs_archive
responses
factory_boy
flake8
pytz
lcogt-logging
django-extensions
ipython
